requirements.txt
matplotlib             
matplotlib-inline      
numpy                  
opencv-python          
opencv-python-headless 
pandas                
pillow                
seaborn                
torch                  
torchaudio             
torchmetrics          
torchvision          # CUDA   
tqdm                   
ipython                
         
